refactor(dev): CTL-529 phase 4 — extract broker router.mjs
Moves the bulk of the broker — "today's broker logic" — out of index.mjs
into router.mjs (~1,400 lines): wake emission + the idempotency cache,
the canonical-aware event readers, the registration + agent handlers, the
three deterministic matchers (pr/ticket/phase lifecycle), shouldSkipEvent,
the Groq prose path, debounce batching, the heartbeat watchdog, and the
processEvent central dispatch. router.mjs imports config + state +
projection + broker-state.mjs; nothing imports it except the index barrel
and (next phase) tailer.mjs.

handleWorkerStateChanged moves here too — it is an event handler dispatched
by processEvent, and housing it here (rather than in projection.mjs)
avoids the projection→router cycle its getEventPayload/getEventOrchestrator
use would otherwise create. It reaches the projected-file writers through
the existing router→projection edge.

Two structural additions, forced by the module split, both behavior-
preserving: getEventName and startWatchdog gain `export` (the tailer and
main consume them); clearDebounceTimers() is a new export so the daemon
shutdown path can stop the now-router-internal debounce timers — it does
exactly the clearTimeout(debounceTimer)/clearTimeout(hardCapTimer) pair
shutdown() ran inline before.

index.mjs drops to 542 lines and re-exports the 29 public router symbols.
Some now-unused imports remain in index.mjs; they are cleaned in phase 5
when the tailer extraction finalizes the thin main.

Broker suite green: 273 pass / 0 fail across 7 files.
